Перенести в презентер доску
